HACK4IMPACT-master/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
chromadb>=0.4.0
Pillow>=10.0.0
pytesseract>=0.3.10
simple-websocket>=1.0.0
groq>=0.11.0
python-dotenv>=1.0.0
